Car Classification API/requirements.txt
flask
werkzeug
gevent
monk-pytorch-cpu
torch==1.5.0
torchvision==0.6.0
matplotlib==3.2.0